requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
flask-socketio
bcrypt
python-dotenv
gunicorn
